% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABCrun2LongPreProccessed}
\alias{ABCrun2LongPreProccessed}
\title{tirefittingr built in datasets}
\format{
A data frame with 606 rows and 8 variables:
\describe{
  \item{SL}{Slip Ratio (unitless)}
  \item{IA}{Inclination Angle in degrees}
  \item{FZ}{Normal Load in N}
  \item{FX}{Longitudinal Load in N}
  \item{SA}{Slip Angle in degrees}
  \item{TSTC}{Temperature in C}
  \item{FY}{Lateral Load in N}
  \item{P}{Pressure in KPa}
}
}
\usage{
ABCrun2LongPreProccessed
}
\description{
Fake tire data made from randomly generated parameters
}
\keyword{datasets}
