% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateRSS}
\alias{calculateRSS}
\title{Calculates RSS Residuals Error between a pacejka Curve and data set}
\usage{
calculateRSS(svInputFile = NULL, parameters)
}
\arguments{
\item{svInputFile}{string vector. Path names or names of data frames}

\item{parameters}{dataframe or numeric vector. Pacejka parameters.}
}
\value{

}
\description{
Calculates RSS Residuals Error between a pacejka Curve and data set
}
