% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clearTirefittingrOptions}
\alias{clearTirefittingrOptions}
\title{Clears all options from tirefittingr}
\usage{
clearTirefittingrOptions()
}
\description{
Clears all options from tirefittingr
}
\examples{
clearTirefittingrOptions()
}
