% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{createFitDataFrame}
\alias{createFitDataFrame}
\title{Creates a data frame with a column}
\usage{
createFitDataFrame(fFittingFunction, sParameterArgName = "parameters", ...)
}
\arguments{
\item{fFittingFunction}{function. A tire fititng function. Pre-installed
functions available are FYPurePacejka2002, FXPurePacejka2002.NoIA, and
FXPurePacejka2002.wIA . The function must have assigned attributes
"parameterNames", which list the names of the parameters.
Example: attr(fFittingFunction, "parameterNames") =
c("pC1", "pD1", "pD2" ...)}

\item{sParameterArgName}{string. Defaults to "parameters".
Name of the parameters argument in the
curve fitting function. For built in curves such as FYPurePacejka2002,
the parameter name is 'parameters'.}

\item{...}{Other arguments passed into the curve fitting function. Pass
vectors in order to sweep through multiple values. The exception being
the item containing the vector of parameters, which  will not be
swept through.}
}
\value{
data frame with a column for each original item from ... and the
  output of the fFittingFunction. This data frame is ready for plotting.
}
\description{
Creates a data frame with a column
}
\examples{
data = createFitDataFrame(
  FYPurePacejka2002,
  sParameterArgName = "parameters",
  SA = seq(-12,12,0.25),
   IA = 0,
  FZ = c(-250, -500, -750, -1000, -1500),
  parameters = c(20,5.89,-1.23,-12.23,1.87,-0.75,0.08,90.44,0,0.02,
  -0.18,187.77,11.81,-23.34,-0.04,-0.14,0.42,0.02))
}
