% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfStartParFX}
\alias{dfStartParFX}
\title{Data Frame of Starting Population for fitting FXPurePacejka2002}
\format{
A data frame
}
\usage{
dfStartParFX
}
\description{
Starting population is 150 sets of randomly generated coefficients that
  provide a starting point for the genetic solver. The minimum and maximum
  limits of each parameter in the solution are taken from the minimum
  and maximum values in this data frame. Population works with both
  FXPurePacejka2002.wIA and FXPurePacejka2002.NoIA.
}
\keyword{datasets}
