% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfStartParFY}
\alias{dfStartParFY}
\title{Data Frame of Starting Population for fitting FYPurePacejka2002}
\format{
A data frame
}
\usage{
dfStartParFY
}
\description{
Starting population is 180 sets of randomly generated coefficients that
  provide a starting point for the genetic solver. The minimum and maximum
  limits of each parameter in the solution are taken from the minimum
  and maximum values in this data frame.
}
\keyword{datasets}
