% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{fFYPlot}
\alias{fFYPlot}
\title{Plotting Function for FY}
\usage{
fFYPlot(dfPlotData, runSummary)
}
\arguments{
\item{dfPlotData}{data frame. Data to plot.}

\item{runSummary}{data frame. From the fitTires function.}
}
\description{
Used by the `fitTires` function. Set with `setFYPure2002()` or manually with
  `options(tirefittingr.sfPlot = "fFYPlot")`.
}
