% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{FXPre}
\alias{FXPre}
\title{Preprocess Data For FX}
\usage{
FXPre(dfData)
}
\arguments{
\item{dfData}{data frame. Data to be processed}
}
\value{
data frame. Processed data.
}
\description{
Depending on the options for
}
