% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Magic.R
\name{FXPurePacejka2002.NoIA}
\alias{FXPurePacejka2002.NoIA}
\title{Pacejka 2002 Pure Longitudinal Magic Formula}
\usage{
FXPurePacejka2002.NoIA(FZ, SL, FZ0 = -1600, Vs = 40, V0 = 40, parameters)
}
\arguments{
\item{FZ}{numeric. Normal Load in Newtons.}

\item{SL}{numeric. Slip Ratio in SAE standard, unitless.
SL is Slip Ratio based on RE (such that SL=0 gives FX=0).
See Race Car Vehicle Dynamics page 62 (Milliken/Milliken).
For TTC Data, use column named "SL" not "SR".}

\item{FZ0}{numeric. Nominal rated load in Newtons. Defaults to -1600N.
Typically the highest load used in the data.
See section 4.3.2 of Tyre and Vehicle Dynamics (Pacejka, 2002)}

\item{Vs}{numeric. Tire Slip velocity in kph. Defaults to 40 kph.}

\item{V0}{numeric. Reference Velocity in kph. Defaults to 40 kph.
Typically the average velocity of the test.}

\item{parameters}{numeric Vector. 14 Pacejka Pure-Longitudinal
coefficients in the following order:
pC1, pD1, pD2, pE1, pE2, pE3, pE4, pK1, pK2, pK3, pV1, pV2,
pH1, pH2}
}
\value{
Longitudinal Force, FX, in Newtons.
}
\description{
Uses the Pure Longitudinal Magic Formula (Pacejka 2002) as written, to
  calculate FX. This does not include variation with inclination angle.
  Various sources include a 15th parameter, pD3, to include IA variation,
  which is not present in this function. To include camber variation, see
  \code{\link{FXPurePacejka2002.wIA}}.
}
\details{
Check out the source code on the tirefittingr github in /R/Magic.R.

Works in the SAE Sign conventions: Fz <0, (See Race Car Vehicle Dynamics
  by Milliken, p39).
  or \url{https://www.oreilly.com/library/view/tire-and-vehicle/9780080970165/images/F150016bm01-9780080970165.jpg}
}
\examples{
params = c(1,3.16,-0.62,0.18,3.58,-10.54,-0.25,-0.01,-0.01,
  -100,40,2.24,-0.1,-0.45)
  FX = FXPurePacejka2002.NoIA(
      FZ = -1000,
      SL = 0.1,
      parameters = params)
  #Often used within createFitDataFrameToPlot function to create a plotable
  #data frame.
  mGridPlot = createFitDataFrame(FXPurePacejka2002.NoIA, "parameters",
  SL = seq(from = -0.2, to = 0.2, by = 0.005),
  FZ = c(c(-250, -750, -1200)),
  parameters = params)

}
\references{
Tyre and Vehicle Dynamics (Pacejka, 2002)
  Race Car Vehicle Dynamics (Milliken)
}
\seealso{
Other MagicFormulas: 
\code{\link{FXPurePacejka2002.wIA}()},
\code{\link{FYPurePacejka2002}()}
}
\concept{MagicFormulas}
\keyword{pacejka}
