% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Magic.R
\name{FXPurePacejka2002.wIA}
\alias{FXPurePacejka2002.wIA}
\alias{FXPureMF52}
\title{MF5.2 / Pacejka 2002 Pure Longitudinal Magic Formula with IA}
\usage{
FXPurePacejka2002.wIA(FZ, SL, IA, FZ0 = -1600, parameters)

FXPureMF52(FZ, SL, IA, FZ0 = -1600, parameters)
}
\arguments{
\item{FZ}{numeric. Normal Load in Newtons.}

\item{SL}{numeric. Slip Ratio in SAE standard, unitless.
SL is Slip Ratio based on RE (such that SL=0 gives FX=0).
See Race Car Vehicle Dynamics page 62 (Milliken/Milliken).
For TTC Data, use column named "SL" not "SR".}

\item{IA}{numeric. Inclination Angle in degrees.
Similar to camber angle, except the positive/negative sign has a
different convention. Caution: While fitting to data, it is highly
recommended that you use a dataset containing a sweep of different IA
values. While you can fit to a dataset that only uses one IA value,
any model created this way will output garbage if any value of
IA is input other than the one it was originally fit at.}

\item{FZ0}{numeric. Nominal rated load in Newtons. Defaults to -1600N.
Typically the highest load used in the data.
See section 4.3.2 of Tyre and Vehicle Dynamics (Pacejka, 2002)}

\item{parameters}{numeric Vector. 15 Pacejka Pure-Longitudinal
coefficients in the following order:
pC1, pD1, pD2, pD3, pE1, pE2, pE3, pE4, pK1, pK2, pK3, pV1, pV2,
pH1, pH2}
}
\value{
Longitudinal Force, FX, in Newtons.
}
\description{
Uses the Pure Longitudinal Magic Formula (Pacejka 2002). \strong{With a
  common modification:} the addition of a \strong{15th parameter, pD3},
  to include IA variation, which was not present in the standard 2002
  pure longitudinal function. For the unmodified function, see
  \code{\link{FXPurePacejka2002.wIA}}. FXPureMF52 is for
  Magic Formula 5.2 Pure longitudinal, which is the same as the Pacejka 2002
  Pure longitudinal Model. FXPureMF52 function is an alias to
  FXPurePacejka2002.wIA.
}
\details{
Check out the source code on the tirefittingr github in /R/Magic.R.

Works in the SAE Sign conventions: Fz <0, (See Race Car Vehicle Dynamics
by Milliken, p39).
  or \url{https://www.oreilly.com/library/view/tire-and-vehicle/
  9780080970165/images/F150016bm01-9780080970165.jpg}
}
\examples{
params = c(1,3.16,-0.62,15.59,0.18,3.58,-10.54,-0.25,-0.01,-0.01,
  -100,40,2.24,-0.1,-0.45)
  FX = FXPurePacejka2002.wIA(
      FZ = -1000,
      SL = 0.1,
      IA = 0,
      parameters = params)
  #Often used within createFitDataFrameToPlot function to create a plotable
  #data frame.
  mGridPlot = createFitDataFrame(FXPurePacejka2002.wIA, "parameters",
  SL = seq(from = -0.2, to = 0.2, by = 0.005), IA = 0,
  FZ = c(c(-250, -750, -1200)),
  parameters = params)
}
\references{
Tyre and Vehicle Dynamics (Pacejka, 2002).
  Race Car Vehicle Dynamics (Milliken)
}
\seealso{
Other MagicFormulas: 
\code{\link{FXPurePacejka2002.NoIA}()},
\code{\link{FYPurePacejka2002}()}

Other MagicFormulas: 
\code{\link{FXPurePacejka2002.NoIA}()},
\code{\link{FYPurePacejka2002}()}
}
\concept{MagicFormulas}
\keyword{pacejka}
