% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{FYPre}
\alias{FYPre}
\title{Pre-Process Data For FY}
\usage{
FYPre(dfData)
}
\arguments{
\item{dfData}{data frame. Data to be processed}
}
\value{
data frame. Processed data.
}
\description{
Filters out cold data depending on the setting for
`options("tirefittingr.coldCutoffTemp")`
}
