% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Magic.R
\name{FYPurePacejka2002}
\alias{FYPurePacejka2002}
\alias{FYPureMF52}
\title{MF5.2 / Pacejka 2002 Pure Lateral Magic Formula}
\usage{
FYPurePacejka2002(
  FZ,
  SA,
  IA,
  Vs = 40,
  FZ0 = -1600,
  V0 = 40,
  parameters,
  scaleMu = 1
)

FYPureMF52(FZ, SA, IA, Vs = 40, FZ0 = -1600, V0 = 40, parameters, scaleMu = 1)
}
\arguments{
\item{FZ}{numeric. Normal Load in Newtons.}

\item{SA}{numeric. Slip Angle in degrees.}

\item{IA}{numeric. Inclination Angle in degrees.
Similar to camber angle, except the positive/negative sign has a
different convention. Caution: While fitting to data, it is highly
recommended that you use a dataset containing a sweep of different IA
values. While you can fit to a dataset that only uses one IA value,
any model created this way will output garbage if any value of
IA is input other than the one it was originally fit at.}

\item{Vs}{numeric. Tire Slip velocity in kph. Defaults to 40 kph.}

\item{FZ0}{numeric. Nominal rated load in Newtons. Defaults to -1600N.
Typically the highest load used in the data.
See section 4.3.2 of Tyre and Vehicle Dynamics (Pacejka, 2002)}

\item{V0}{numeric. Reference Velocity in kph. Defaults to 40 kph.
Typically the average velocity of the test.}

\item{parameters}{numeric Vector. 18 Pacejka Pure-Lateral
coefficients in the following order:
pC1, pD1, pD2, pD3, pE1, pE2, pE3, pE4, pK1, pK2, pK3, pV1, pV2,
pV3, pV4, pH1, pH2, pH3}

\item{scaleMu}{numeric. Defaults to 1. Scaling factor multiplied by mu.
Used as symbol 'lambda_mu_y' in Tyre and Vehicle Dynamics (Pacejka, 2002)
Set to 1 during fitting. Testing from a facility (eg. TTC) is
done on a sandpaper "road" which has a higher coefficient of friction
(mu) than regular pavement. To use this function to estimate real forces
that would be seen on pavement (for lapsim, etc), mu must be scaled
down to what can be expected from a road. Typical mu on sandpaper may be
2.5, while pavement may be 1.4 (roughly equal to max steady state lateral
Gs without Aerodynamic downforce). As such a scaling factor of
scaleMu = 1.6/2.5 = 0.64 may be used so that the output will be reasonable
for a tire on pavement.}
}
\value{
Lateral Force, FY, in Newtons.
}
\description{
Calculates the lateral tire force given FZ, SA, IA, and a set of parameters
  using the Pacejke 2002 pure lateral formula. FYPureMF52 is for
  Magic Formula 5.2 Pure lateral, which is the same as the Pacejka 2002 Pure
  Lateral Model. FYPureMF52 function is an alias to FYPurePacejka2002.
}
\details{
Check out the source code on the tirefittingr github in /R/Magic.R.

Works in the SAE Sign conventions: Fz <0, (See Race Car Vehicle Dynamics
by Milliken, p39).
  or \url{https://www.oreilly.com/library/view/tire-and-vehicle/9780080970165/images/F150016bm01-9780080970165.jpg}
}
\examples{
 params = c(20,5.89,-1.23,-12.23,1.87,-0.75,0.08,90.44,0,0.02,
  -0.18,187.77,11.81,-23.34,-0.04,-0.14,0.42,0.02)
  FY = FYPurePacejka2002(
      FZ = -1000,
      SA = 10,
      IA = 0,
      parameters = params)
  #Often used within createFitDataFrameToPlot function to create a plotable
  #data frame.
  mGridPlot = createFitDataFrame(FYPurePacejka2002, "parameters",
  SA = seq(from = -12, to = 12, by = 0.1), IA = 0,
  FZ = c(c(-250, -750, -1200)),
  parameters = params)
}
\references{
Tyre and Vehicle Dynamics (Pacejka, 2002)
}
\seealso{
\code{\link{FXPurePacejka2002.NoIA}}

Other MagicFormulas: 
\code{\link{FXPurePacejka2002.NoIA}()},
\code{\link{FXPurePacejka2002.wIA}()}

Other MagicFormulas: 
\code{\link{FXPurePacejka2002.NoIA}()},
\code{\link{FXPurePacejka2002.wIA}()}
}
\concept{MagicFormulas}
\keyword{pacejka}
