% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{plotGoodnessOfFit}
\alias{plotGoodnessOfFit}
\title{Compare Fitted Curve to Raw Data}
\usage{
plotGoodnessOfFit(dfPlotData, lCurveParameters, lPlotSetup)
}
\arguments{
\item{dfPlotData}{data frame. Contains raw data to plot.}

\item{lCurveParameters}{list containing the parameters for the curves.
Any values containing NA will be silently dropped.}

\item{lPlotSetup}{list containing the following named items: \describe{
\item{fFitFunction}{string. Contains the name of a magic formula such as
FYPurePacejka2002.}
\item{iDataPointsPlot}{integer. Number of raw data points to plot.}
\item{sInCol}{Name of the column of raw data that will be plotted on the
x axis.}
\item{sOutCol}{Name of the column of raw data that will be plotted on the
y axis.}
\item{sSweep}{Name of one column of raw data, as well as an argument in
the magic formula that will be swept through a range of values.}
\item{lConst}{Named list of inputs to the magic formulas that will be
kept constant. Such as list(IA = 0)}
\item{xlab}{x axis label}
\item{ylab}{y axis label}}}
}
\description{
Plots raw data with a fitted curve overlayed so that the user can judge the
  quality of the fit.
}
\details{
UNTESTED OUTSIDE OF fitTires function!
}
