% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{readCSVTireData}
\alias{readCSVTireData}
\title{Reads TTC .dat Data Files}
\usage{
readCSVTireData(sRunFileName)
}
\arguments{
\item{sRunFileName}{string. Complete path of the file name of a raw
data file, or the name of a dataframe variable already loaded into memory
Note: forward slashes "/"
must be used in R in file paths, not backslashes "\".}
}
\value{
data frame with named columns containing raw data.
}
\description{
A Wrapper for \code{\link{readTireData}}
  Runs \code{\link{readTireData}} with `iColNamesRow = 1`,
  `iFirstDataRow = 2`, and `sep = ","`.
  `return(readTireData(sRunFileName, iColNamesRow = 1, iFirstDataRow = 2, sep =","))`
}
\examples{
  \dontrun{
  readCSVTireData("C:/folder/full_path_to_file.csv") #use forward slashes.
  }
}
\seealso{
Other Reading Tire Data Functions: 
\code{\link{readTTCData}()},
\code{\link{readTireData}()}
}
\concept{Reading Tire Data Functions}
