% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{readTireData}
\alias{readTireData}
\title{Reads Tire Data Files}
\usage{
readTireData(sRunFileName, iColNamesRow = 1, iFirstDataRow = 2, sep = "\\t")
}
\arguments{
\item{sRunFileName}{string. Complete path of the file name of a raw
data file, or the name of a dataframe variable already loaded into memory
Note: forward slashes "/"
must be used in R in file paths, not backslashes "\".}

\item{iColNamesRow}{integer. The row number of the row that contains
the column names. These column names will be the names of the output
data frame.}

\item{iFirstDataRow}{integer. The row number of the first row that contains
data.}

\item{sep}{string. The delimiting character used in the text file, such as
"," for a csv file. Use "\\t" for a tab delimited file.
See \link[utils]{read.table} for more details.}
}
\value{
data frame with named columns containing raw data.
}
\description{
Opens a Tire Data file, optionally dropping rows that do not contain data
  such as title row and units row.
}
\examples{
data = readTTCData("C:/users/data.dat")
}
\seealso{
Other Reading Tire Data Functions: 
\code{\link{readCSVTireData}()},
\code{\link{readTTCData}()}
}
\concept{Reading Tire Data Functions}
