% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postprocess.R
\name{removeSimilarValues}
\alias{removeSimilarValues}
\title{Removes Duplicates}
\usage{
removeSimilarValues(x, dWithin)
}
\arguments{
\item{x}{numeric vector}

\item{dWithin}{numeric. How close will values be to trigger averaging
the two values into one?}
}
\value{
numeric vector where the values close enough to be duplicates
  are removed by averaging them together.
}
\description{
Removes duplicates from a list to post-process peaks from
  `IDPmisc::peaks` function. Removes duplicates by averaging 2 values
  that are within the `dWithin` value.
}
