% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{setFXPure2002.wIA}
\alias{setFXPure2002.wIA}
\alias{setFXMF52}
\title{Changes Options to Run FXPure.wIA}
\usage{
setFXPure2002.wIA()

setFXMF52()
}
\description{
Sets the following options:\preformatted{  options(tirefittingr.sdfStartPop = "dfStartParFX",
    tirefittingr.sfFittingFunction = "FXPurePacejka2002.wIA",
    tirefittingr.sfPreProcess = "FXPre",
    tirefittingr.sfPlot = "fFXPlot")
}

So that \code{fitTires()} can be run for \code{FXPure2002.wIA}. For all
available options, see the help file for the \code{\link{fitTires}}
function.
}
\examples{
setFXPure2002.wIA()
}
