% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateRSS}
\alias{calculateRSS}
\title{Calculates RSS Residuals Error Between a Pacejka Curve and Data Set}
\usage{
calculateRSS(svInputFile = NULL, parameters)
}
\arguments{
\item{svInputFile}{string vector. Path names or names of data frames}

\item{parameters}{dataframe or numeric vector. Pacejka parameters.}
}
\value{
data frame. First column is RSS divided by number of datapoints.
  Second column is number of data points.
}
\description{
Calculates RSS divided by the total number of data points in the full
  data set. See \url{https://en.wikipedia.org/wiki/Residual_sum_of_squares}
  {https://en.wikipedia.org/wiki/Residual_sum_of_squares}.
}
