% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{setFXPure2002.NoIA}
\alias{setFXPure2002.NoIA}
\title{Changes Options to Run \code{FXPure.NoIA}}
\usage{
setFXPure2002.NoIA()
}
\description{
Sets the following options:\preformatted{  options(tirefittingr.sdfStartPop = "dfStartParFX",
    tirefittingr.sfFittingFunction = "FXPurePacejka2002.NoIA",
    tirefittingr.sfPreProcess = "FXPre",
    tirefittingr.sfPlot = "fFXPlot.NoIA")
}

So that \code{fitTires} function can be run for \code{FXPure2002.NoIA}. For all
available options, see the help file for the \code{\link{fitTires}}
function.
}
\examples{
setFXPure2002.NoIA()
}
